Move "external" tests from test/external/ to externaltests/ and rename them .spec.js.
This makes it possible to get jest to execute them.
Also, those files will be subject to linting.
