+++ externaltests/deepObject.spec.js
@@ -1,4 +1,4 @@
-var expect = require('../../');
+var expect = require('../');
 it('should not dot out the object in the diff', function () {
     expect(function () {
         expect({0:{0:{0:{0:{0:{0:{}}}}}}}, 'to equal', {});
