+++ externaltests/forgotToReturnPendingPromiseFromFailingItBlock.spec.js
@@ -1,4 +1,4 @@
-var expect = require('../../');
+var expect = require('../');
 
 it('should call the callback', function () {
     expect(function () {}, 'to call the callback');
