+++ externaltests/notToErrorCaughtError.jasmine.json
@@ -1,0 +1,6 @@
+{
+  "spec_dir": "externaltests",
+  "spec_files": [
+    "notToErrorCaughtError.spec.js"
+  ]
+}
