+++ externaltests/notToErrorCaughtError.spec.js
@@ -1,4 +1,4 @@
-var expect = require('../../');
+var expect = require('../');
 
 it('should foo', function () {
     return expect(function () {
