--- test/external/fullTrace.externaljs
@@ -1,5 +1,0 @@
-var expect = require('../../').clone().use(require('./node_modules/unexpected-bogus/index.externaljs'));
-
-it('should fail with a full trace', function () {
-    expect('abc', 'to be bogus');
-});
