--- test/external/notToErrorCaughtError.jasmine.json
@@ -1,6 +1,0 @@
-{
-  "spec_dir": "test/external",
-  "spec_files": [
-    "notToErrorCaughtError.externaljs"
-  ]
-}
